requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==5.0.1
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.11.2
billiard==4.2.1
bleach==6.0.0
celery==5.5.3
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.0.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.1.2
configobj==5.0.9
contourpy==1.0.7
cryptography==45.0.6
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.1
dvc==3.62.0
dvc-data==3.16.11
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
ecdsa==0.19.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==1.2.0
fastapi==0.116.1
fastjsonschema==2.16.2
filelock==3.13.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
grandalf==0.8
gto==1.7.2
gunicorn==23.0.0
h11==0.16.0
htmlmin==0.1.12
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.3.2
idna==3.4
ImageHash==4.3.1
iniconfig==2.1.0
ipykernel==6.21.2
ipympl==0.9.3
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
iterative-telemetry==0.0.10
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.1
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.6.4
multimethod==1.9.1
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nltk==3.9.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
omegaconf==2.3.0
orjson==3.11.2
outcome==1.3.0.post0
packaging==23.0
pandas==1.5.3
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==4.3.8
pluggy==1.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.37
propcache==0.3.2
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.6.1
pycparser==2.21
pydantic==1.10.5
pydot==4.0.1
pygit2==1.18.2
Pygments==2.14.0
pygtrie==2.5.0
pyparsing==3.2.3
pyrsistent==0.19.3
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.5.0
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==305; sys_platform == "win32"
pywinpty==2.0.10; sys_platform == "win32"
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2025.7.34
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.1
scipy==1.9.3
scmrepo==3.5.2
seaborn==0.12.2
selenium==4.28.1
semver==3.0.4
Send2Trash==1.8.0
shellingham==1.5.4
shiboken6==6.5.2
shortuuid==1.0.13
shtab==1.7.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqltrie==0.11.2
stack-data==0.6.2
starlette==0.47.2
statsmodels==0.13.5
sympy==1.12
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
terminado==0.17.1
threadpoolctl==3.6.0
tinycss2==1.2.1
tomli==2.2.1
tomlkit==0.13.3
torch==2.1.1
torchvision==0.16.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trio==0.28.0
trio-websocket==0.11.1
typeguard==2.13.3
typer==0.16.0
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.35.0
vine==5.1.0
visions==0.7.5
voluptuous==0.15.2
wcwidth==0.2.6
webcolors==1.12
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.5
wsproto==1.2.0
yarl==1.20.1
ydata-profiling==4.0.0
zc.lockfile==3.0.post1
